Support multiple docker networks
The iptables script currently assumes a single docker network, which it creates rules for. But if there are multiple, the `DOCKER_CIDR` variable will have newline-separated ips, which breaks the iptables rules. It turns out that iptables will accept comma-separated ips, so we just need to join all the extracted ips with a comma.